requirements.txt
discord.py
python-aiml